[refactor] Code componentization related to absent schedules
